move monotonicity definition to header file
